backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]~=0.31.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
azure-ai-projects>=2.0.0b3
azure-identity~=1.19.0
azure-cosmos~=4.7.0
cryptography~=42.0.0
yfinance~=0.2.36
duckduckgo-search~=7.0.0
agent-framework-core>=1.0.0rc2
agent-framework-anthropic>=1.0.0b260304
azure-ai-agents